requirements.txt
Django==5.1.5
gunicorn==21.2.0
psycopg2-binary==2.9.10
dj-database-url==2.1.0
whitenoise==6.6.0
python-dotenv==1.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-extensions==3.2.3
Pillow==10.4.0